+++ crates/fuel-core/src/query/balance/asset_query.rs
@@ -1,4 +1,5 @@
 use crate::graphql_api::database::ReadView;
+use fuel_core_services::stream::IntoBoxStream;
 use fuel_core_storage::{
     iter::IterDirection,
     Error as StorageError,
@@ -14,19 +15,20 @@ use fuel_core_types::{
         AssetId,
     },
 };
-use itertools::Itertools;
+use futures::Stream;
 use std::collections::HashSet;
+use tokio_stream::StreamExt;
 
 /// At least required `target` of the query per asset's `id` with `max` coins.
 #[derive(Clone)]
 pub struct AssetSpendTarget {
     pub id: AssetId,
     pub target: u64,
-    pub max: usize,
+    pub max: u16,
 }
 
 impl AssetSpendTarget {
-    pub fn new(id: AssetId, target: u64, max: usize) -> Self {
+    pub fn new(id: AssetId, target: u64, max: u16) -> Self {
         Self { id, target, max }
     }
 }
@@ -48,9 +50,10 @@ impl Exclude {
     }
 }
 
+#[derive(Clone)]
 pub struct AssetsQuery<'a> {
     pub owner: &'a Address,
-    pub assets: Option<HashSet<&'a AssetId>>,
+    pub allowed_assets: Option<HashSet<&'a AssetId>>,
     pub exclude: Option<&'a Exclude>,
     pub database: &'a ReadView,
     pub base_asset_id: &'a AssetId,
@@ -59,27 +62,32 @@ pub struct AssetsQuery<'a> {
 impl<'a> AssetsQuery<'a> {
     pub fn new(
         owner: &'a Address,
-        assets: Option<HashSet<&'a AssetId>>,
+        allowed_assets: Option<HashSet<&'a AssetId>>,
         exclude: Option<&'a Exclude>,
         database: &'a ReadView,
         base_asset_id: &'a AssetId,
     ) -> Self {
         Self {
             owner,
-            assets,
+            allowed_assets,
             exclude,
             database,
             base_asset_id,
         }
     }
 
-    fn coins_iter(&self) -> impl Iterator<Item = StorageResult<CoinType>> + '_ {
+    fn coins_iter(mut self) -> impl Stream<Item = StorageResult<CoinType>> + 'a {
+        let allowed_assets = self.allowed_assets.take();
         self.database
             .owned_coins_ids(self.owner, None, IterDirection::Forward)
             .map(|id| id.map(CoinId::from))
-            .filter_ok(|id| {
-                if let Some(exclude) = self.exclude {
-                    !exclude.coin_ids.contains(id)
+            .filter(move |result| {
+                if let Ok(id) = result {
+                    if let Some(exclude) = self.exclude {
+                        !exclude.coin_ids.contains(id)
+                    } else {
+                        true
+                    }
                 } else {
                     true
                 }
@@ -91,27 +99,34 @@ impl<'a> AssetsQuery<'a> {
                     } else {
                         return Err(anyhow::anyhow!("The coin is not UTXO").into());
                     };
+                    // TODO: Fetch coin in a separate thread
                     let coin = self.database.coin(id)?;
 
                     Ok(CoinType::Coin(coin))
                 })
             })
-            .filter_ok(|coin| {
-                if let CoinType::Coin(coin) = coin {
-                    self.has_asset(&coin.asset_id)
+            .filter(move |result| {
+                if let Ok(CoinType::Coin(coin)) = result {
+                    allowed_asset(&allowed_assets, &coin.asset_id)
                 } else {
                     true
                 }
             })
     }
 
-    fn messages_iter(&self) -> impl Iterator<Item = StorageResult<CoinType>> + '_ {
+    fn messages_iter(&self) -> impl Stream<Item = StorageResult<CoinType>> + 'a {
+        let exclude = self.exclude;
+        let database = self.database;
         self.database
             .owned_message_ids(self.owner, None, IterDirection::Forward)
             .map(|id| id.map(CoinId::from))
-            .filter_ok(|id| {
-                if let Some(exclude) = self.exclude {
-                    !exclude.coin_ids.contains(id)
+            .filter(move |result| {
+                if let Ok(id) = result {
+                    if let Some(e) = exclude {
+                        !e.coin_ids.contains(id)
+                    } else {
+                        true
+                    }
                 } else {
                     true
                 }
@@ -123,11 +138,18 @@ impl<'a> AssetsQuery<'a> {
                     } else {
                         return Err(anyhow::anyhow!("The coin is not a message").into());
                     };
-                    let message = self.database.message(&id)?;
+                    // TODO: Fetch message in a separate thread (https://github.com/FuelLabs/fuel-core/pull/2340)
+                    let message = database.message(&id)?;
                     Ok(message)
                 })
             })
-            .filter_ok(|message| message.data().is_empty())
+            .filter(|result| {
+                if let Ok(message) = result {
+                    message.data().is_empty()
+                } else {
+                    true
+                }
+            })
             .map(|result| {
                 result.map(|message| {
                     CoinType::MessageCoin(
@@ -139,28 +161,23 @@ impl<'a> AssetsQuery<'a> {
             })
     }
 
-    fn has_asset(&self, asset_id: &AssetId) -> bool {
-        self.assets
-            .as_ref()
-            .map(|assets| assets.contains(asset_id))
-            .unwrap_or(true)
-    }
-
     /// Returns the iterator over all valid(spendable, allowed by `exclude`) coins of the `owner`.
     ///
     /// # Note: The coins of different type are not grouped by the `asset_id`.
     // TODO: Optimize this by creating an index
     //  https://github.com/FuelLabs/fuel-core/issues/588
-    pub fn coins(&self) -> impl Iterator<Item = StorageResult<CoinType>> + '_ {
-        let has_base_asset = self.has_asset(self.base_asset_id);
-        let messages_iter = has_base_asset
-            .then(|| self.messages_iter())
-            .into_iter()
-            .flatten();
-        self.coins_iter().chain(messages_iter)
+    pub fn coins(self) -> impl Stream<Item = StorageResult<CoinType>> + 'a {
+        let has_base_asset = allowed_asset(&self.allowed_assets, self.base_asset_id);
+        if has_base_asset {
+            let message_iter = self.messages_iter();
+            self.coins_iter().chain(message_iter).into_boxed_ref()
+        } else {
+            self.coins_iter().into_boxed_ref()
+        }
     }
 }
 
+#[derive(Clone)]
 pub struct AssetQuery<'a> {
     pub owner: &'a Address,
     pub asset: &'a AssetSpendTarget,
@@ -196,7 +213,14 @@ impl<'a> AssetQuery<'a> {
 
     /// Returns the iterator over all valid(spendable, allowed by `exclude`) coins of the `owner`
     /// for the `asset_id`.
-    pub fn coins(&self) -> impl Iterator<Item = StorageResult<CoinType>> + '_ {
+    pub fn coins(self) -> impl Stream<Item = StorageResult<CoinType>> + 'a {
         self.query.coins()
     }
+}
+
+fn allowed_asset(allowed_assets: &Option<HashSet<&AssetId>>, asset_id: &AssetId) -> bool {
+    allowed_assets
+        .as_ref()
+        .map(|allowed_assets| allowed_assets.contains(asset_id))
+        .unwrap_or(true)
 }