+++ crates/fuel-core/src/schema/block.rs
@@ -7,9 +7,9 @@ use crate::{
     fuel_core_graphql_api::{
         api_service::ConsensusModule,
         database::ReadView,
+        query_costs,
         Config as GraphQLConfig,
         IntoApiResult,
-        QUERY_COSTS,
     },
     schema::{
         scalars::{
@@ -36,11 +36,7 @@ use async_graphql::{
     Union,
 };
 use fuel_core_storage::{
-    iter::{
-        BoxedIter,
-        IntoBoxedIter,
-        IterDirection,
-    },
+    iter::IterDirection,
     Result as StorageResult,
 };
 use fuel_core_types::{
@@ -50,6 +46,10 @@ use fuel_core_types::{
     },
     fuel_types,
     fuel_types::BlockHeight,
+};
+use futures::{
+    Stream,
+    StreamExt,
 };
 
 pub struct Block(pub(crate) CompressedBlock);
@@ -112,14 +112,14 @@ impl Block {
         self.0.header().clone().into()
     }
 
-    #[graphql(complexity = "QUERY_COSTS.storage_read + child_complexity")]
+    #[graphql(complexity = "query_costs().storage_read + child_complexity")]
     async fn consensus(&self, ctx: &Context<'_>) -> async_graphql::Result<Consensus> {
         let query = ctx.read_view()?;
         let height = self.0.header().height();
         Ok(query.consensus(height)?.try_into()?)
     }
 
-    #[graphql(complexity = "QUERY_COSTS.block_transactions_ids")]
+    #[graphql(complexity = "query_costs().block_transactions_ids")]
     async fn transaction_ids(&self) -> Vec<TransactionId> {
         self.0
             .transactions()
@@ -129,7 +129,7 @@ impl Block {
     }
 
     // Assume that in average we have 32 transactions per block.
-    #[graphql(complexity = "QUERY_COSTS.block_transactions + child_complexity")]
+    #[graphql(complexity = "query_costs().block_transactions + child_complexity")]
     async fn transactions(
         &self,
         ctx: &Context<'_>,
@@ -240,7 +240,7 @@ pub struct BlockQuery;
 
 #[Object]
 impl BlockQuery {
-    #[graphql(complexity = "QUERY_COSTS.block_header + child_complexity")]
+    #[graphql(complexity = "query_costs().block_header + child_complexity")]
     async fn block(
         &self,
         ctx: &Context<'_>,
@@ -270,7 +270,7 @@ impl BlockQuery {
     }
 
     #[graphql(complexity = "{\
-        (QUERY_COSTS.block_header + child_complexity) \
+        (query_costs().block_header + child_complexity) \
         * (first.unwrap_or_default() as usize + last.unwrap_or_default() as usize) \
     }")]
     async fn blocks(
@@ -298,7 +298,7 @@ pub struct HeaderQuery;
 
 #[Object]
 impl HeaderQuery {
-    #[graphql(complexity = "QUERY_COSTS.block_header + child_complexity")]
+    #[graphql(complexity = "query_costs().block_header + child_complexity")]
     async fn header(
         &self,
         ctx: &Context<'_>,
@@ -312,7 +312,7 @@ impl HeaderQuery {
     }
 
     #[graphql(complexity = "{\
-        (QUERY_COSTS.block_header + child_complexity) \
+        (query_costs().block_header + child_complexity) \
         * (first.unwrap_or_default() as usize + last.unwrap_or_default() as usize) \
     }")]
     async fn headers(
@@ -339,16 +339,14 @@ fn blocks_query<T>(
     query: &ReadView,
     height: Option<BlockHeight>,
     direction: IterDirection,
-) -> BoxedIter<StorageResult<(U32, T)>>
+) -> impl Stream<Item = StorageResult<(U32, T)>> + '_
 where
     T: async_graphql::OutputType,
     T: From<CompressedBlock>,
 {
-    let blocks = query.compressed_blocks(height, direction).map(|result| {
+    query.compressed_blocks(height, direction).map(|result| {
         result.map(|block| ((*block.header().height()).into(), block.into()))
-    });
-
-    blocks.into_boxed()
+    })
 }
 
 #[derive(Default)]