+++ crates/fuel-core/src/schema/gas_price.rs
@@ -5,7 +5,7 @@ use super::scalars::{
 use crate::{
     graphql_api::{
         api_service::GasPriceProvider,
-        QUERY_COSTS,
+        query_costs,
     },
     schema::ReadViewProvider,
 };
@@ -42,7 +42,7 @@ pub struct LatestGasPriceQuery {}
 
 #[Object]
 impl LatestGasPriceQuery {
-    #[graphql(complexity = "QUERY_COSTS.block_header")]
+    #[graphql(complexity = "query_costs().block_header")]
     async fn latest_gas_price(
         &self,
         ctx: &Context<'_>,
@@ -80,7 +80,7 @@ pub struct EstimateGasPriceQuery {}
 
 #[Object]
 impl EstimateGasPriceQuery {
-    #[graphql(complexity = "2 * QUERY_COSTS.storage_read")]
+    #[graphql(complexity = "2 * query_costs().storage_read")]
     async fn estimate_gas_price(
         &self,
         ctx: &Context<'_>,