+++ crates/fuel-core/src/schema/upgrades.rs
@@ -1,8 +1,8 @@
 use crate::{
     graphql_api::{
         api_service::ConsensusProvider,
+        query_costs,
         IntoApiResult,
-        QUERY_COSTS,
     },
     schema::{
         chain::ConsensusParameters,
@@ -29,7 +29,7 @@ pub struct UpgradeQuery;
 
 #[Object]
 impl UpgradeQuery {
-    #[graphql(complexity = "QUERY_COSTS.storage_read + child_complexity")]
+    #[graphql(complexity = "query_costs().storage_read + child_complexity")]
     async fn consensus_parameters(
         &self,
         ctx: &Context<'_>,
@@ -42,7 +42,7 @@ impl UpgradeQuery {
         Ok(ConsensusParameters(params))
     }
 
-    #[graphql(complexity = "QUERY_COSTS.storage_read + child_complexity")]
+    #[graphql(complexity = "query_costs().storage_read + child_complexity")]
     async fn state_transition_bytecode_by_version(
         &self,
         ctx: &Context<'_>,
@@ -54,7 +54,7 @@ impl UpgradeQuery {
             .into_api_result()
     }
 
-    #[graphql(complexity = "QUERY_COSTS.storage_read + child_complexity")]
+    #[graphql(complexity = "query_costs().storage_read + child_complexity")]
     async fn state_transition_bytecode_by_root(
         &self,
         root: HexString,
@@ -73,7 +73,7 @@ impl StateTransitionBytecode {
         HexString(self.root.to_vec())
     }
 
-    #[graphql(complexity = "QUERY_COSTS.state_transition_bytecode_read")]
+    #[graphql(complexity = "query_costs().state_transition_bytecode_read")]
     async fn bytecode(
         &self,
         ctx: &Context<'_>,